achordion: allow chording of thumb keys and non-thumb keys
The Achordion library rejects chording of keys on the same half to avoid
accidentally triggering e.g. home row mods when rolling during typing.

The default logic does not consider thumb keys separately. However, this
makes it hard to trigger keys on a layer when the key is on the same
hand as the thumb key used for layer-tap. Since accidental overlaps
between thumb keys and non-thumb keys are infrequent (probably less
frequently overlap between the left and right hands), it makes more
sense to consider the thumb keys to be separate from non-thumb keys for
the chording check.

This change overrides the chording check in this manner for the Voyager.